requirements.txt
# Core (required)
pyyaml>=6.0

# Parquet export (optional but recommended)
pyarrow>=12.0

# S3 export (optional)
boto3>=1.28
